Restructure plot files
